Resolve #240 - Rearrange character stats, add more stats
